Added infantry & reinforcements to eastern objectives + minor changes
Added a bunch of infantry squads and snipers around the eastern objectives.
Placed down four LSVs, a pawnee and a GMG hunter that are called in as support once the fighting starts.
Replaced the engineering GMG hunter with a HMG hunter.
Added insignias to the guys at base
Added AT launchers and killzone markers to Thronos.
Probably some other stuff.